Add whole bunch of logging and properly control racing during startup.

+++ src/rustc_daemon/mod.rs
@@ -1,3 +1,5 @@
+use std::fs::OpenOptions;
+use std::hash::{DefaultHasher, Hash, Hasher};
 use std::io::{Read, Write};
 use std::num::NonZeroUsize;
 use std::os::unix::net::{UnixListener, UnixStream};
@@ -21,7 +23,7 @@ use rustc_session::{EarlyDiagCtxt, config};
 use rustc_span::source_map::SourceMap;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Hash)]
 struct CompileTask {
     args: Vec<String>,
     env: Vec<(String, String)>,
@@ -45,17 +47,28 @@ fn invoke_daemon() -> ExitCode {
     let early_dcx = EarlyDiagCtxt::new(ErrorOutputType::default());
     let args = args::raw_args(&early_dcx);
 
+    let env: Vec<_> = env::vars().collect();
+    let working_directory = env::current_dir().expect("Unable to get current directory");
+
+    let task = CompileTask { args: args.clone(), env, working_directory };
+
+    let mut h = DefaultHasher::new();
+    task.hash(&mut h);
+    let hash = h.finish();
+
+    eprintln!("I: Start ({hash}) as PID {}", std::process::id());
+
     let at_args = args.get(1..).unwrap_or_default();
 
     let early_args = args::arg_expand_all(&early_dcx, at_args);
 
     match handle_options(&early_dcx, &early_args) {
-        HandledOptions::None => ExitCode::SUCCESS,
+        HandledOptions::None => {
+            eprintln!("I: Finished in handle options ({hash})");
+            ExitCode::SUCCESS
+        }
         HandledOptions::Normal(matches) if !matches.opt_present("print") => {
-            let env: Vec<_> = env::vars().collect();
-            let working_directory = env::current_dir().expect("Unable to get current directory");
-
-            let task = CompileTask { args, env, working_directory };
+            eprintln!("I: Connecting ({hash})");
 
             let socket_path = socket_path();
             let mut retrying = false;
@@ -63,8 +76,9 @@ fn invoke_daemon() -> ExitCode {
                 match UnixStream::connect(&socket_path) {
                     Ok(session) => break session,
                     Err(error) => {
+                        eprintln!("I: Could not connect initially ({hash}): {error}");
                         if retrying {
-                            println!("Cannot connect to daemon: {error}");
+                            eprintln!("I: Cannot connect to daemon ({hash}): {error}");
                             return ExitCode::FAILURE;
                         } else {
                             retrying = true;
@@ -74,26 +88,33 @@ fn invoke_daemon() -> ExitCode {
                 }
             };
 
+            eprintln!("I: Starting request ({hash})");
             serde_json::to_writer(&mut session, &task).expect("Unable to send task to daemon");
             session
                 .shutdown(std::net::Shutdown::Write)
                 .expect("Could not shutdown write connection");
+            eprintln!("I: Sent request ({hash})");
 
-            let _ = std::io::copy(&mut session, &mut std::io::stderr());
+            let e = std::io::copy(&mut session, &mut std::io::stderr());
+
+            eprintln!("I: End of request ({hash}): {e:?}");
 
             ExitCode::SUCCESS
         }
         _ => {
+            eprintln!("I: Handle internally ({hash})");
             init_rustc_env_logger(&early_dcx);
             install_ice_hook(DEFAULT_BUG_REPORT_URL, |_| ());
 
             catch_with_exit_code(|| run_compiler(&args, &mut NoopCallbacks));
+            eprintln!("I: Finished internally ({hash})");
             ExitCode::SUCCESS
         }
     }
 }
 
 fn start_daemon() {
+    eprintln!("Started daemon");
     let directory =
         tempfile::tempdir().expect("Unable to get temporary directory for notification socket");
     let mut notify_path = directory.path().to_owned();
@@ -109,6 +130,7 @@ fn start_daemon() {
     let (mut socket, _) = listener.accept().expect("Unable to head start of daemon");
     let mut status = vec![];
     socket.read_to_end(&mut status).expect("Unable to read startup status");
+    eprintln!("Received startup notification: {status:?}");
 }
 
 struct DaemonCallbacks {
@@ -152,6 +174,8 @@ pub unsafe fn discard_inherited_jobserver() {
 }
 
 fn run_daemon() -> ExitCode {
+    eprintln!("D: Starting as PID {}", std::process::id());
+
     // We need to discard the parent's job server otherwise we could keep it around for unrelated compilation sessions.
     // Safe because nothing is going on yet.
     unsafe {
@@ -179,8 +203,23 @@ fn run_daemon() -> ExitCode {
         .expect("Unable to become a daemon");
 
     let socket_path = socket_path();
-    let _ = std::fs::remove_file(&socket_path);
-    let listener = UnixListener::bind(socket_path);
+    let lock_path = lock_path();
+
+    let listener = {
+        let lock_file = OpenOptions::new()
+            .read(true)
+            .write(true)
+            .create(true)
+            .open(lock_path)
+            .expect("Could not open lockfile.");
+        lock_file.lock().expect("Unable to lock lockfile");
+        if UnixStream::connect(&socket_path).is_ok() {
+            Err(std::io::ErrorKind::AddrInUse.into())
+        } else {
+            let _ = std::fs::remove_file(&socket_path);
+            UnixListener::bind(socket_path)
+        }
+    };
 
     let early_dcx = EarlyDiagCtxt::new(ErrorOutputType::default());
 
@@ -197,16 +236,28 @@ fn run_daemon() -> ExitCode {
     if let Ok(path) = env::var("__CG_CLIF_DAEMON_NOTIFY") {
         let mut notify = UnixStream::connect(path).expect("Unable to notify startup");
         notify.write_all(b"Started").expect("Unable to notify startup");
+
+        eprintln!("Notified startup");
     }
 
-    let listener = listener.expect("Unable to open socket");
+    let Ok(listener) = listener else {
+        eprintln!("Already running");
+        return ExitCode::SUCCESS;
+    };
 
     loop {
         let token = jobserver::client().acquire().expect("Unable to acquire jobserver token");
         match listener.accept() {
             Ok((mut stream, _)) => {
                 std::thread::spawn(move || {
+                    let mut hash = 0;
+
+                    eprintln!("D: Received connection ({hash})");
                     if let Ok(request) = serde_json::from_reader::<_, CompileTask>(&mut stream) {
+                        let mut h = DefaultHasher::new();
+                        request.hash(&mut h);
+                        hash = h.finish();
+                        eprintln!("D: Request ({hash})");
                         catch_with_exit_code(|| {
                             run_compiler(
                                 &request.args,
@@ -217,7 +268,10 @@ fn run_daemon() -> ExitCode {
                                 },
                             )
                         });
+                    } else {
+                        eprintln!("D: Error receiving request ({hash})");
                     }
+                    eprintln!("D: Job done ({hash})");
                     // Explicit token to keep track that we are running a job.
                     drop(token);
                 });
@@ -325,4 +379,12 @@ fn socket_path() -> PathBuf {
     socket_path.push("cg_clif_daemon");
 
     socket_path
+}
+
+fn lock_path() -> PathBuf {
+    let mut lock_path: PathBuf =
+        env::var_os("XDG_RUNTIME_DIR").expect("Missing runtime directory").into();
+    lock_path.push("cg_clif_daemon_lock");
+
+    lock_path
 }
